Split up `number_format.dart`.
PiperOrigin-RevId: 328335245
